debug contollers

+++ src/controllers/user.controller.js
@@ -1,4 +1,5 @@
 const User = require("../models/user.model.js");
+const Wird = require("../models/wird.model.js");
 
 async function getUser(req, res) {
   try {
@@ -24,6 +25,16 @@ async function deleteUser(req, res) {
     if (!telegramid) {
       return res.status(400).json({ message: "Missing telegramid header" });
     }
+
+    const wirds = await Wird.deleteMany({ "owner.telegramId": telegramid });
+    if (!wirds) {
+      return res
+        .status(404)
+        .json({
+          message: "Wirds by the user with this telegram id are not found.",
+        });
+    }
+
     const user = await User.findOneAndDelete({ telegramId: telegramid });
     if (!user) {
       return res.status(404).json({ message: "Telegram user not found" });
